feat(site): real chart-viz deck on profile (dark recap band)
